backend/requirements.txt
fastapi==0.109.0
uvicorn==0.24.0
httpx==0.26.0
pydantic==2.5.3
email-validator==2.1.1
python-dotenv==1.0.1
openai==2.14.0
